Replace Alpine analyzer module with Syft call (#697)
* refactor syft handlers module layout


* add alpine syft handler function


* migrate handler dispatch to handlers module


* update alpine test defaults + remove syft analyzer shebang statement


* bump syft to v0.5.1
