+++ tests/functional/clients/standalone/package_list/test_alpine.py
@@ -1,7 +1,6 @@
 import sys
 from .fixtures import alpine
 from . import path_params
-import json
 import pytest
 
 
@@ -23,7 +22,7 @@ class TestAlpinePaths:
     def test_pkgs_allinfo(self, analyzed_data, pkg, metadata):
         result = analyzed_data("py38")
         pkgs = result['image']['imagedata']['analysis_report']['package_list']['pkgs.allinfo']['base']
-        loaded = dict(json.loads(pkgs.get(pkg, '{}')))
+        loaded = dict(pkgs.get(pkg, {}))
         
         # a separate test exists just for the files attribute, leave this out of the assertion
         loaded.pop('files')
@@ -36,7 +35,7 @@ class TestAlpinePaths:
     def test_pkgs_allinfo_files(self, analyzed_data, pkg, metadata):
         result = analyzed_data("py38")
         pkgs = result['image']['imagedata']['analysis_report']['package_list']['pkgs.allinfo']['base']
-        loaded = json.loads(pkgs.get(pkg, '{}'))
+        loaded = pkgs.get(pkg, {})
         assert sorted(loaded['files']) == sorted(metadata['files'])
 
     @pytest.mark.parametrize('pkg,version', [pytest.param(pkg, version, id=pkg) for pkg, version in alpine.pkgs_plus_source_all.items()])
